[components][wip] move translation logic to mixin base class
