+++ python_modules/dagster/dagster/components/utils/__init__.py
@@ -1,0 +1,94 @@
+import importlib.util
+import subprocess
+import sys
+import textwrap
+from collections.abc import Iterator
+from contextlib import contextmanager
+from pathlib import Path
+from types import ModuleType
+from typing import TypeVar
+
+import click
+from dagster_shared.error import DagsterError
+
+from dagster import _check as check
+from dagster.components.utils.translation import TranslatorResolvingInfo as TranslatorResolvingInfo
+
+T = TypeVar("T")
+
+
+def exit_with_error(error_message: str) -> None:
+    click.echo(click.style(error_message, fg="red"))
+    sys.exit(1)
+
+
+def format_error_message(message: str) -> str:
+    # width=10000 unwraps any hardwrapping
+    dedented = textwrap.dedent(message).strip()
+    paragraphs = [textwrap.fill(p, width=10000) for p in dedented.split("\n\n")]
+    return "\n\n".join(paragraphs)
+
+
+# Temporarily places a path at the front of sys.path, ensuring that any modules in that path are
+# importable.
+@contextmanager
+def ensure_loadable_path(path: Path) -> Iterator[None]:
+    orig_path = sys.path.copy()
+    sys.path.insert(0, str(path))
+    try:
+        yield
+    finally:
+        sys.path = orig_path
+
+
+def get_path_for_package(package_name: str) -> str:
+    spec = importlib.util.find_spec(package_name)
+    if not spec:
+        raise DagsterError(f"Cannot find package: {package_name}")
+    # file_path = spec.origin
+    submodule_search_locations = spec.submodule_search_locations
+    if not submodule_search_locations:
+        raise DagsterError(f"Package does not have any locations for submodules: {package_name}")
+    return submodule_search_locations[0]
+
+
+# ########################
+# ##### PLATFORM
+# ########################
+
+
+def is_windows() -> bool:
+    return sys.platform == "win32"
+
+
+def is_macos() -> bool:
+    return sys.platform == "darwin"
+
+
+# ########################
+# ##### VENV
+# ########################
+
+
+def get_venv_executable(venv_dir: Path, executable: str = "python") -> Path:
+    if is_windows():
+        return venv_dir / "Scripts" / f"{executable}.exe"
+    else:
+        return venv_dir / "bin" / executable
+
+
+def install_to_venv(venv_dir: Path, install_args: list[str]) -> None:
+    executable = get_venv_executable(venv_dir)
+    command = ["uv", "pip", "install", "--python", str(executable), *install_args]
+    subprocess.run(command, check=True)
+
+
+def get_path_from_module(module: ModuleType) -> Path:
+    module_path = (
+        Path(module.__file__).parent
+        if module.__file__
+        else Path(module.__path__[0])
+        if module.__path__
+        else None
+    )
+    return check.not_none(module_path, f"Module {module.__name__} has no filepath")
