+++ python_modules/libraries/dagster-airbyte/dagster_airbyte/components/workspace_component/component.py
@@ -1,14 +1,13 @@
-from collections.abc import Sequence
+from collections.abc import Mapping, Sequence
 from functools import cached_property
-from typing import Annotated, Callable, Optional, Union
+from typing import Annotated, Any, Callable, Optional, Union
 
 import dagster as dg
 import pydantic
 from dagster._core.definitions.job_definition import default_job_io_manager
 from dagster.components.resolved.base import resolve_fields
-from dagster.components.utils import TranslatorResolvingInfo
+from dagster.components.utils.translation import TranslationFn, TranslatorResolvable
 from dagster_shared import check
-from typing_extensions import TypeAlias
 
 from dagster_airbyte.asset_defs import build_airbyte_assets_definitions
 from dagster_airbyte.components.workspace_component.scaffolder import (
@@ -22,34 +21,8 @@ from dagster_airbyte.translator import (
 )
 
 
-def resolve_translation(context: dg.ResolutionContext, model):
-    info = TranslatorResolvingInfo(
-        "props",
-        asset_attributes=model,
-        resolution_context=context,
-        model_key="translation",
-    )
-    return lambda base_asset_spec, props: info.get_asset_spec(
-        base_asset_spec,
-        {
-            "props": props,
-            "spec": base_asset_spec,
-        },
-    )
-
-
-TranslationFn: TypeAlias = Callable[[dg.AssetSpec, AirbyteConnectionTableProps], dg.AssetSpec]
-ResolvedTranslationFn: TypeAlias = Annotated[
-    TranslationFn,
-    dg.Resolver(
-        resolve_translation,
-        model_field_type=Union[str, dg.AssetAttributesModel],
-    ),
-]
-
-
 class ProxyDagsterAirbyteTranslator(DagsterAirbyteTranslator):
-    def __init__(self, fn: TranslationFn):
+    def __init__(self, fn: TranslationFn[AirbyteConnectionTableProps]):
         self.fn = fn
 
     def get_asset_spec(self, props: AirbyteConnectionTableProps) -> dg.AssetSpec:
@@ -98,7 +71,9 @@ def resolve_connection_selector(
 
 
 @dg.scaffold_with(AirbyteCloudWorkspaceComponentScaffolder)
-class AirbyteCloudWorkspaceComponent(dg.Component, dg.Model, dg.Resolvable):
+class AirbyteCloudWorkspaceComponent(
+    dg.Component, dg.Model, TranslatorResolvable[AirbyteConnectionTableProps]
+):
     """Loads Airbyte Cloud connections from a given Airbyte Cloud workspace as Dagster assets.
     Materializing these assets will trigger a sync of the Airbyte Cloud connection, enabling
     you to schedule Airbyte Cloud syncs using Dagster.
@@ -122,10 +97,16 @@ class AirbyteCloudWorkspaceComponent(dg.Component, dg.Model, dg.Resolvable):
             description="Function used to select Airbyte Cloud connections to pull into Dagster.",
         ),
     ] = None
-    translation: Optional[ResolvedTranslationFn] = pydantic.Field(
-        None,
-        description="Function used to translate Airbyte Cloud connection table properties into Dagster asset specs.",
-    )
+
+    @classmethod
+    def get_template_vars_for_translation(
+        cls, data: AirbyteConnectionTableProps
+    ) -> Mapping[str, Any]:
+        return {"props": data}
+
+    @classmethod
+    def get_translator_field_description(cls) -> Optional[str]:
+        return "Function used to translate Airbyte Cloud connection table properties into Dagster asset specs."
 
     @cached_property
     def workspace_resource(self) -> AirbyteCloudWorkspace:
