+++ python_modules/libraries/dagster-dbt/dagster_dbt/components/dbt_project/component.py
@@ -2,22 +2,20 @@ from collections.abc import Iterator, Mapping
 from dataclasses import dataclass
 from functools import cached_property
 from pathlib import Path
-from typing import Annotated, Any, Callable, Optional, Union
+from typing import Annotated, Any, Optional, Union
 
 from dagster import Resolvable
 from dagster._core.definitions.asset_key import AssetKey
-from dagster._core.definitions.assets.definition.asset_spec import AssetSpec
 from dagster._core.definitions.definitions_class import Definitions
 from dagster._core.execution.context.asset_execution_context import AssetExecutionContext
 from dagster._utils.cached_method import cached_method
 from dagster.components.component.component import Component
 from dagster.components.core.context import ComponentLoadContext
 from dagster.components.core.tree import ComponentTree
-from dagster.components.resolved.core_models import AssetAttributesModel, OpSpec, ResolutionContext
+from dagster.components.resolved.core_models import OpSpec, ResolutionContext
 from dagster.components.resolved.model import Resolver
 from dagster.components.scaffold.scaffold import scaffold_with
-from dagster.components.utils import TranslatorResolvingInfo
-from typing_extensions import TypeAlias
+from dagster.components.utils.translation import TranslationFn, TranslatorResolvable
 
 from dagster_dbt.asset_decorator import dbt_assets
 from dagster_dbt.asset_utils import DBT_DEFAULT_EXCLUDE, DBT_DEFAULT_SELECT, get_node
@@ -33,40 +31,12 @@ from dagster_dbt.dbt_manifest_asset_selection import DbtManifestAssetSelection
 from dagster_dbt.dbt_project import DbtProject
 from dagster_dbt.utils import ASSET_RESOURCE_TYPES
 
-TranslationFn: TypeAlias = Callable[[AssetSpec, Mapping[str, Any]], AssetSpec]
-
 
 @dataclass(frozen=True)
 class DagsterDbtComponentsTranslatorSettings(DagsterDbtTranslatorSettings):
     """Subclass of DagsterDbtTranslatorSettings that enables code references by default."""
 
     enable_code_references: bool = True
-
-
-def resolve_translation(context: ResolutionContext, model):
-    info = TranslatorResolvingInfo(
-        "node",
-        asset_attributes=model,
-        resolution_context=context,
-        model_key="translation",
-    )
-    return lambda base_asset_spec, dbt_props: info.get_asset_spec(
-        base_asset_spec,
-        {
-            "node": dbt_props,
-            "spec": base_asset_spec,
-        },
-    )
-
-
-ResolvedTranslationFn: TypeAlias = Annotated[
-    TranslationFn,
-    Resolver(
-        resolve_translation,
-        inject_before_resolve=False,
-        model_field_type=Union[str, AssetAttributesModel],
-    ),
-]
 
 
 @dataclass
@@ -109,7 +79,7 @@ def resolve_dbt_project(context: ResolutionContext, model) -> DbtProject:
 
 @scaffold_with(DbtProjectComponentScaffolder)
 @dataclass
-class DbtProjectComponent(Component, Resolvable):
+class DbtProjectComponent(Component, TranslatorResolvable[Mapping[str, Any]]):
     """Expose a DBT project to Dagster as a set of assets.
 
     This component assumes that you have already set up a dbt project. [Jaffle shop](https://github.com/dbt-labs/jaffle-shop) is their pre-existing
@@ -157,17 +127,6 @@ class DbtProjectComponent(Component, Resolvable):
             ],
         ),
     ] = None
-    translation: Annotated[
-        Optional[ResolvedTranslationFn],
-        Resolver.default(
-            description="Defines how AssetSpecs are translated from dbt nodes",
-            examples=[
-                {
-                    "key": "target/main/{{ node.name }}",
-                },
-            ],
-        ),
-    ] = None
     select: Annotated[
         str,
         Resolver.default(
@@ -199,6 +158,14 @@ class DbtProjectComponent(Component, Resolvable):
             description="Whether to prepare the dbt project every time in `dagster dev` or `dg` cli calls."
         ),
     ] = True
+
+    @classmethod
+    def get_template_vars_for_translation(cls, data: Mapping[str, Any]) -> Mapping[str, Any]:
+        return {"node": data}
+
+    @classmethod
+    def get_translator_field_description(cls) -> Optional[str]:
+        return "Function used to translate dbt nodes into Dagster asset specs."
 
     @cached_property
     def translator(self):
@@ -276,7 +243,9 @@ class DbtProjectComponent(Component, Resolvable):
 class ProxyDagsterDbtTranslator(DagsterDbtTranslator):
     # get_description conflicts on Component, so cant make it directly a translator
 
-    def __init__(self, fn: TranslationFn, settings: Optional[DagsterDbtTranslatorSettings]):
+    def __init__(
+        self, fn: TranslationFn[Mapping[str, Any]], settings: Optional[DagsterDbtTranslatorSettings]
+    ):
         self._fn = fn
         super().__init__(settings)
