+++ python_modules/libraries/dagster-dlt/dagster_dlt/components/dlt_load_collection/component.py
@@ -1,19 +1,17 @@
 import importlib
-from collections.abc import Iterator, Sequence
+from collections.abc import Iterator, Mapping, Sequence
 from dataclasses import dataclass
 from functools import cached_property
-from typing import TYPE_CHECKING, Annotated, Callable, Optional, Union
+from typing import TYPE_CHECKING, Annotated, Any, Callable, Optional
 
 import dagster as dg
 from dagster import AssetKey, AssetSpec, Component, ComponentLoadContext, Resolvable, Resolver
 from dagster._core.execution.context.asset_execution_context import AssetExecutionContext
 from dagster.components.resolved.context import ResolutionContext
-from dagster.components.resolved.core_models import AssetAttributesModel
 from dagster.components.scaffold.scaffold import scaffold_with
-from dagster.components.utils import TranslatorResolvingInfo
+from dagster.components.utils.translation import TranslationFn, TranslatorResolvable
 from dlt import Pipeline
 from dlt.extract.source import DltSource
-from typing_extensions import TypeAlias
 
 from dagster_dlt.asset_decorator import dlt_assets
 from dagster_dlt.components.dlt_load_collection.scaffolder import DltLoadCollectionScaffolder
@@ -21,8 +19,6 @@ from dagster_dlt.translator import DagsterDltTranslator, DltResourceTranslatorDa
 
 if TYPE_CHECKING:
     from dagster_dlt import DagsterDltResource
-
-TranslationFn: TypeAlias = Callable[[AssetSpec, DltResourceTranslatorData], AssetSpec]
 
 
 def _load_object_from_python_path(resolution_context: ResolutionContext, path: str):
@@ -44,7 +40,7 @@ def _load_object_from_python_path(resolution_context: ResolutionContext, path: s
 class ComponentDagsterDltTranslator(DagsterDltTranslator):
     """Custom base translator, which generates keys from dataset and table names."""
 
-    def __init__(self, *, fn: Optional[TranslationFn] = None):
+    def __init__(self, *, fn: Optional[TranslationFn[DltResourceTranslatorData]] = None):
         super().__init__()
         self._fn = fn or (lambda spec, _: spec)
 
@@ -62,35 +58,8 @@ class ComponentDagsterDltTranslator(DagsterDltTranslator):
         return self._fn(base_asset_spec, data)
 
 
-def resolve_translation(context: ResolutionContext, model):
-    info = TranslatorResolvingInfo(
-        "data",
-        asset_attributes=model,
-        resolution_context=context,
-        model_key="translation",
-    )
-    return lambda base_asset_spec, data: info.get_asset_spec(
-        base_asset_spec,
-        {
-            "resource": data.resource,
-            "pipeline": data.pipeline,
-            "spec": base_asset_spec,
-        },
-    )
-
-
-ResolvedTranslationFn: TypeAlias = Annotated[
-    TranslationFn,
-    Resolver(
-        resolve_translation,
-        inject_before_resolve=False,
-        model_field_type=Union[str, AssetAttributesModel],
-    ),
-]
-
-
 @dataclass
-class DltLoadSpecModel(Resolvable):
+class DltLoadSpecModel(TranslatorResolvable[DltResourceTranslatorData]):
     """Represents a single dlt load, a combination of pipeline and source."""
 
     pipeline: Annotated[
@@ -104,7 +73,12 @@ class DltLoadSpecModel(Resolvable):
             model_field_type=str,
         ),
     ]
-    translation: Optional[ResolvedTranslationFn] = None
+
+    @classmethod
+    def get_template_vars_for_translation(
+        cls, data: DltResourceTranslatorData
+    ) -> Mapping[str, Any]:
+        return {"resource": data.resource, "pipeline": data.pipeline}
 
     @cached_property
     def translator(self):
